Move to trait so they share same logic
