+++ app/Models/Score/Best/Model.php
@@ -20,9 +20,7 @@ use DB;
  */
 abstract class Model extends BaseModel implements Traits\ReportableInterface
 {
-    use Traits\Reportable, Traits\WithDbCursorHelper;
-
-    public ?float $weight = null;
+    use Traits\Reportable, Traits\WithDbCursorHelper, Traits\WithWeightedPp;
 
     protected $macros = [
         'accurateRankCounts',
@@ -65,11 +63,6 @@ abstract class Model extends BaseModel implements Traits\ReportableInterface
         }
 
         return null;
-    }
-
-    public function weightedPp(): ?float
-    {
-        return $this->weight === null ? null : $this->weight * $this->pp;
     }
 
     public function macroForListing()
